+++ src/app/ApplicationNodeImplementation.cpp
@@ -48,7 +48,7 @@ namespace viscom {
             tex.emplace_back(static_cast<GLenum>(gl::GLenum::GL_RGBA32F));
         }
 
-        debugTextureBuffers_ = CreateOffscreenBuffers(FrameBufferDescriptor{tex, std::vector<RenderBufferDescriptor>{}});
+        debugTextureBuffers_ = CreateOffscreenBuffers(FrameBufferDescriptor{tex,{}});
         freeCam_->SetCameraPosition(glm::vec3(0,1,8));
         for(const auto& uniform : quad_->GetUniforms()) {
             LOG(INFO) << uniform.first << ": " << glbinding::Meta::getString(uniform.second.type);
@@ -156,21 +156,8 @@ namespace viscom {
     void ApplicationNodeImplementation::DrawFrame(FrameBuffer& fbo)
     {
         SelectOffscreenBuffer(debugTextureBuffers_)->DrawToFBO([this](){
-            gfx::Box b;
-            b.position = glm::vec3(0);
-            b.bounds = glm::vec3(.25f);
 
-            auto position = GetCamera()->GetPosition();
-            auto camOrientation = (glm::mat3_cast(GetCamera()->GetOrientation()));
-            auto MVP = (GetCamera()->GetViewPerspectiveMatrix());
-            auto prog = quad_->GetGPUProgram();
             {
-                gl::glUseProgram(prog->getProgramId());
-                gl::glUniform1f(prog->getUniformLocation("u_time"), time_);
-                gl::glUniformMatrix4fv(prog->getUniformLocation("u_MVP"), 1, gl::GL_FALSE, glm::value_ptr(MVP));
-                gl::glUniform3f(prog->getUniformLocation("u_camPosition"), position.x, position.y, position.z);
-                quad_->SendSubroutines();
-                b.send(prog, 0);
                 quad_->Draw();
             }
         });
